fix the "sortah hangs when redirecting to another router" bug
Had to do with misuse of throw/rescue instead of throw/catch. This commit also adds an acceptance test to ensure that similar problems won't pop up. This also makes CleanRoom have to be a non-basic object. Not sure how to fix this yet. Finally, the FinishedExecution exception is no longer there, the throw/catch machinery doesn't need it.